azure ping and test fix
