+++ runtime/drivers/azure/azure.go
@@ -3,12 +3,20 @@ package azure
 import (
 	"context"
 	"errors"
+	"fmt"
+	"net/url"
+	"os"
 
+	"github.com/Azure/azure-sdk-for-go/sdk/azidentity"
+	"github.com/Azure/azure-sdk-for-go/sdk/storage/azblob"
+	"github.com/Azure/azure-sdk-for-go/sdk/storage/azblob/container"
 	"github.com/mitchellh/mapstructure"
 	"github.com/rilldata/rill/runtime/drivers"
 	"github.com/rilldata/rill/runtime/pkg/activity"
+	"github.com/rilldata/rill/runtime/pkg/blob"
 	"github.com/rilldata/rill/runtime/storage"
 	"go.uber.org/zap"
+	"gocloud.dev/blob/azureblob"
 )
 
 func init() {
@@ -72,6 +80,7 @@ type ConfigProperties struct {
 	Key              string `mapstructure:"azure_storage_key"`
 	SASToken         string `mapstructure:"azure_storage_sas_token"`
 	ConnectionString string `mapstructure:"azure_storage_connection_string"`
+	Bucket           string `mapstructure:"azure_storage_bucket"`
 	AllowHostAccess  bool   `mapstructure:"allow_host_access"`
 }
 
@@ -116,7 +125,23 @@ var _ drivers.Handle = &Connection{}
 
 // Ping implements drivers.Handle.
 func (c *Connection) Ping(ctx context.Context) error {
-	return drivers.ErrNotImplemented
+	if c.config.Bucket == "" {
+		return fmt.Errorf("bucket not configured")
+	}
+
+	// Try to open the bucket to verify connection
+	bucket, err := c.openBucket(ctx, c.config.Bucket, false)
+	if err != nil {
+		return fmt.Errorf("failed to open bucket: %w", err)
+	}
+	defer bucket.Close()
+
+	_, err = bucket.ListObjects(ctx, "*")
+	if err != nil {
+		return fmt.Errorf("failed to list objects: %w", err)
+	}
+
+	return nil
 }
 
 // Driver implements drivers.Connection.
@@ -203,4 +228,147 @@ func (c *Connection) AsWarehouse() (drivers.Warehouse, bool) {
 // AsNotifier implements drivers.Connection.
 func (c *Connection) AsNotifier(properties map[string]any) (drivers.Notifier, error) {
 	return nil, drivers.ErrNotNotifier
+}
+
+// newClient returns a new azure blob client.
+func (c *Connection) newClient(bucket string) (*container.Client, error) {
+	var accountKey, sasToken, connectionString string
+
+	accountName, err := c.accountName()
+	if err != nil {
+		return nil, err
+	}
+
+	if c.config.AllowHostAccess {
+		accountKey = os.Getenv("AZURE_STORAGE_KEY")
+		sasToken = os.Getenv("AZURE_STORAGE_SAS_TOKEN")
+		connectionString = os.Getenv("AZURE_STORAGE_CONNECTION_STRING")
+	}
+
+	if c.config.Key != "" {
+		accountKey = c.config.Key
+	}
+	if c.config.SASToken != "" {
+		sasToken = c.config.SASToken
+	}
+	if c.config.ConnectionString != "" {
+		connectionString = c.config.ConnectionString
+	}
+
+	if connectionString != "" {
+		client, err := container.NewClientFromConnectionString(connectionString, bucket, nil)
+		if err != nil {
+			return nil, fmt.Errorf("failed container.NewClientFromConnectionString: %w", err)
+		}
+		return client, nil
+	}
+
+	if accountName != "" {
+		svcURL := fmt.Sprintf("https://%s.blob.core.windows.net", accountName)
+		containerURL, err := url.JoinPath(svcURL, bucket)
+		if err != nil {
+			return nil, err
+		}
+
+		var sharedKeyCred *azblob.SharedKeyCredential
+
+		if accountKey != "" {
+			sharedKeyCred, err = azblob.NewSharedKeyCredential(accountName, accountKey)
+			if err != nil {
+				return nil, fmt.Errorf("failed azblob.NewSharedKeyCredential: %w", err)
+			}
+
+			client, err := container.NewClientWithSharedKeyCredential(containerURL, sharedKeyCred, nil)
+			if err != nil {
+				return nil, fmt.Errorf("failed container.NewClientWithSharedKeyCredential: %w", err)
+			}
+			return client, nil
+		}
+
+		if sasToken != "" {
+			serviceURL, err := azureblob.NewServiceURL(&azureblob.ServiceURLOptions{
+				AccountName: accountName,
+				SASToken:    sasToken,
+			})
+			if err != nil {
+				return nil, err
+			}
+
+			containerURL, err := url.JoinPath(string(serviceURL), bucket)
+			if err != nil {
+				return nil, err
+			}
+
+			client, err := container.NewClientWithNoCredential(containerURL, nil)
+			if err != nil {
+				return nil, fmt.Errorf("failed container.NewClientWithNoCredential: %w", err)
+			}
+			return client, nil
+		}
+
+		cred, err := azidentity.NewDefaultAzureCredential(&azidentity.DefaultAzureCredentialOptions{
+			DisableInstanceDiscovery: true,
+		})
+		if err != nil {
+			return nil, fmt.Errorf("failed azidentity.NewDefaultAzureCredential: %w", err)
+		}
+		client, err := container.NewClient(containerURL, cred, nil)
+		if err != nil {
+			return nil, fmt.Errorf("failed container.NewClient: %w", err)
+		}
+		return client, nil
+	}
+
+	return nil, errors.New("can't access remote host without credentials: no credentials provided")
+}
+
+func (c *Connection) newAnonymousClient(bucket string) (*container.Client, error) {
+	accountName, err := c.accountName()
+	if err != nil {
+		return nil, err
+	}
+
+	svcURL := fmt.Sprintf("https://%s.blob.core.windows.net", accountName)
+	containerURL, err := url.JoinPath(svcURL, bucket)
+	if err != nil {
+		return nil, err
+	}
+	client, err := container.NewClientWithNoCredential(containerURL, nil)
+	if err != nil {
+		return nil, fmt.Errorf("failed container.NewClientWithNoCredential: %w", err)
+	}
+
+	return client, nil
+}
+
+func (c *Connection) accountName() (string, error) {
+	if c.config.Account != "" {
+		return c.config.Account, nil
+	}
+
+	if c.config.AllowHostAccess {
+		return os.Getenv("AZURE_STORAGE_ACCOUNT"), nil
+	}
+
+	return "", errors.New("account name not found")
+}
+
+func (c *Connection) openBucket(ctx context.Context, bucket string, anonymous bool) (*blob.Bucket, error) {
+	var client *container.Client
+	var err error
+	if anonymous {
+		client, err = c.newAnonymousClient(bucket)
+	} else {
+		client, err = c.newClient(bucket)
+	}
+	if err != nil {
+		return nil, err
+	}
+
+	azureBucket, err := azureblob.OpenBucket(ctx, client, nil)
+	if err != nil {
+		return nil, err
+	}
+
+	return blob.NewBucket(azureBucket, c.logger)
 }
