+++ runtime/drivers/azure/object_store.go
@@ -4,16 +4,10 @@ import (
 	"context"
 	"errors"
 	"fmt"
-	"net/url"
-	"os"
 
-	"github.com/Azure/azure-sdk-for-go/sdk/azidentity"
-	"github.com/Azure/azure-sdk-for-go/sdk/storage/azblob"
-	"github.com/Azure/azure-sdk-for-go/sdk/storage/azblob/container"
 	"github.com/rilldata/rill/runtime/drivers"
 	"github.com/rilldata/rill/runtime/pkg/blob"
 	"github.com/rilldata/rill/runtime/pkg/globutil"
-	"gocloud.dev/blob/azureblob"
 )
 
 // ListObjects implements drivers.ObjectStore.
@@ -59,153 +53,18 @@ func (c *Connection) DownloadFiles(ctx context.Context, path string) (drivers.Fi
 func (c *Connection) parseBucketURL(path string) (*globutil.URL, error) {
 	url, err := globutil.ParseBucketURL(path)
 	if err != nil {
+		if errors.Is(err, globutil.ErrMissingScheme) {
+			// Missing schema means its only path and use the default bucket from connector config
+			return &globutil.URL{
+				Scheme: "azure",
+				Host:   c.config.Bucket,
+				Path:   path,
+			}, nil
+		}
 		return nil, fmt.Errorf("failed to parse path %q: %w", path, err)
 	}
 	if url.Scheme != "az" && url.Scheme != "azure" {
 		return nil, fmt.Errorf("invalid Azure path %q: should start with az://", path)
 	}
 	return url, nil
-}
-
-func (c *Connection) openBucket(ctx context.Context, bucket string, anonymous bool) (*blob.Bucket, error) {
-	var client *container.Client
-	var err error
-	if anonymous {
-		client, err = c.newAnonymousClient(bucket)
-	} else {
-		client, err = c.newClient(bucket)
-	}
-	if err != nil {
-		return nil, err
-	}
-
-	azureBucket, err := azureblob.OpenBucket(ctx, client, nil)
-	if err != nil {
-		return nil, err
-	}
-
-	return blob.NewBucket(azureBucket, c.logger)
-}
-
-// newClient returns a new azure blob client.
-func (c *Connection) newClient(bucket string) (*container.Client, error) {
-	var accountKey, sasToken, connectionString string
-
-	accountName, err := c.accountName()
-	if err != nil {
-		return nil, err
-	}
-
-	if c.config.AllowHostAccess {
-		accountKey = os.Getenv("AZURE_STORAGE_KEY")
-		sasToken = os.Getenv("AZURE_STORAGE_SAS_TOKEN")
-		connectionString = os.Getenv("AZURE_STORAGE_CONNECTION_STRING")
-	}
-
-	if c.config.Key != "" {
-		accountKey = c.config.Key
-	}
-	if c.config.SASToken != "" {
-		sasToken = c.config.SASToken
-	}
-	if c.config.ConnectionString != "" {
-		connectionString = c.config.ConnectionString
-	}
-
-	if connectionString != "" {
-		client, err := container.NewClientFromConnectionString(connectionString, bucket, nil)
-		if err != nil {
-			return nil, fmt.Errorf("failed container.NewClientFromConnectionString: %w", err)
-		}
-		return client, nil
-	}
-
-	if accountName != "" {
-		svcURL := fmt.Sprintf("https://%s.blob.core.windows.net", accountName)
-		containerURL, err := url.JoinPath(svcURL, bucket)
-		if err != nil {
-			return nil, err
-		}
-
-		var sharedKeyCred *azblob.SharedKeyCredential
-
-		if accountKey != "" {
-			sharedKeyCred, err = azblob.NewSharedKeyCredential(accountName, accountKey)
-			if err != nil {
-				return nil, fmt.Errorf("failed azblob.NewSharedKeyCredential: %w", err)
-			}
-
-			client, err := container.NewClientWithSharedKeyCredential(containerURL, sharedKeyCred, nil)
-			if err != nil {
-				return nil, fmt.Errorf("failed container.NewClientWithSharedKeyCredential: %w", err)
-			}
-			return client, nil
-		}
-
-		if sasToken != "" {
-			serviceURL, err := azureblob.NewServiceURL(&azureblob.ServiceURLOptions{
-				AccountName: accountName,
-				SASToken:    sasToken,
-			})
-			if err != nil {
-				return nil, err
-			}
-
-			containerURL, err := url.JoinPath(string(serviceURL), bucket)
-			if err != nil {
-				return nil, err
-			}
-
-			client, err := container.NewClientWithNoCredential(containerURL, nil)
-			if err != nil {
-				return nil, fmt.Errorf("failed container.NewClientWithNoCredential: %w", err)
-			}
-			return client, nil
-		}
-
-		cred, err := azidentity.NewDefaultAzureCredential(&azidentity.DefaultAzureCredentialOptions{
-			DisableInstanceDiscovery: true,
-		})
-		if err != nil {
-			return nil, fmt.Errorf("failed azidentity.NewDefaultAzureCredential: %w", err)
-		}
-		client, err := container.NewClient(containerURL, cred, nil)
-		if err != nil {
-			return nil, fmt.Errorf("failed container.NewClient: %w", err)
-		}
-		return client, nil
-	}
-
-	return nil, errors.New("can't access remote host without credentials: no credentials provided")
-}
-
-func (c *Connection) newAnonymousClient(bucket string) (*container.Client, error) {
-	accountName, err := c.accountName()
-	if err != nil {
-		return nil, err
-	}
-
-	svcURL := fmt.Sprintf("https://%s.blob.core.windows.net", accountName)
-	containerURL, err := url.JoinPath(svcURL, bucket)
-	if err != nil {
-		return nil, err
-	}
-	client, err := container.NewClientWithNoCredential(containerURL, nil)
-	if err != nil {
-		return nil, fmt.Errorf("failed container.NewClientWithNoCredential: %w", err)
-	}
-
-	return client, nil
-}
-
-func (c *Connection) accountName() (string, error) {
-	if c.config.Account != "" {
-		return c.config.Account, nil
-	}
-
-	if c.config.AllowHostAccess {
-		return os.Getenv("AZURE_STORAGE_ACCOUNT"), nil
-	}
-
-	return "", errors.New("account name not found")
 }
