+++ runtime/drivers/duckdb/model_executor_sqlstore_self_mysql_test.go
@@ -99,7 +99,7 @@ func TestMySQLToDuckDBTransfer(t *testing.T) {
 	require.NoError(t, err)
 
 	t.Run("model_executor_mysql_to_duckDB", func(t *testing.T) {
-		mysqlToDuckDB(t, fmt.Sprintf("host=%s port=%v database=mydb user=myuser password=mypassword", host, port.Int()))
+		mysqlToDuckDB(t, dsn)
 	})
 }
