+++ runtime/pkg/globutil/globutil.go
@@ -1,9 +1,12 @@
 package globutil
 
 import (
+	"errors"
 	"fmt"
 	"strings"
 )
+
+var ErrMissingScheme = errors.New("Missing URL Scheme")
 
 type URL struct {
 	Scheme string
@@ -33,7 +36,7 @@ func (u *URL) String() string {
 func ParseBucketURL(path string) (*URL, error) {
 	scheme, path, ok := strings.Cut(path, "://")
 	if !ok {
-		return nil, fmt.Errorf("failed to parse URL '%q'", path)
+		return nil, fmt.Errorf("%w: failed to parse URL '%s'", ErrMissingScheme, path)
 	}
 
 	host, path, ok := strings.Cut(path, "/")
